ADV Codespace only/requirements.txt
audio_recorder_streamlit==0.0.8
googletrans-py
llama_index==0.9.14.post1
nest_asyncio==1.5.8
numpy==1.26.2
pydub==0.25.1
python-dotenv==1.0.0
SpeechRecognition==3.10.1
streamlit==1.29.0
openai
sentence-transformers
pypdf
google-generativeai
